getting closer to having a single form that can both encrypt and decrypt
	The idea here is for users to be able to store data in a remote
database, but with client side AES encryption. I create a new address
for each website that I sign up to and with each I have an address
and a password, (or openID). I already store the website location and
some notes against each alias, but don't store the password as I
don't want unencrypted passwords in my database.
  That said, I want my users to have the option of not encrypting passwords
if they chose not to enter a key. It would be nice for the users to be able
to see the cypher-text, (hence the Encrypt button) but I don't think that
it is needed.